FIX: __main__ 파일을 실행하는 구조로 변경
