+++ chanel/__init__.py
@@ -1,25 +1,1 @@
-import datetime
-
-from sanic import Sanic
-from sanic_jwt_extended import JWTManager
-
-from chanel.config import SERVICE_NAME, LOGO
-from chanel.config.setting import SETTINGS
-from chanel.config.vault import VaultClient
-from chanel.config.listener import initialize
-from chanel.controller import api
-
-
-def create_app() -> Sanic:
-    _app = Sanic(SERVICE_NAME)
-    _app.config.LOGO = LOGO
-    _app.register_listener(initialize, "before_server_start")
-
-    _app.config["JWT_SECRET_KEY"] = SETTINGS.vault_client.get_jwt_secret_key
-    _app.config["JWT_ACCESS_TOKEN_EXPIRES"] = datetime.timedelta(days=1)
-    _app.config["RBAC_ENABLE"] = True
-    JWTManager(_app)
-
-    _app.blueprint(api)
-
-    return _app
+__version__ = "0.0.1"
